test have been rewritten
